Crate runic_types

Modules

capabilities

outputs

Structs

BufWriter

InvalidConversionError

SerializableRecord
A serializable version of log::Record.

Tensor
A multidimensional array with copy-on-write semantics.

TensorView
An immutable view into a Tensor with a particular rank (number of dimensions).

Enums

Type

Value
A dynamically typed value that may be passed back and forth across the runtime.

Traits

AsType
A Rust primitive which has a corresponding Type and can be converted to or from a Value.

HasOutputs

Sink
A consumer of data.

Source
A stream of data, typically something like a random number generator or sensor.

Transform
Process some data, transforming it from one form to another.