Trait runic_types::AsType

pub trait AsType where
    Self: Into<Value>,
    Self: TryFrom<Value>, {
    const TYPE: Type;
}

A Rust primitive which has a corresponding Type and can be converted to or from a Value.

Associated Constants

const TYPE: Type

The corresponding Type variant.

Loading content...